Extract public API module
